Added textAlign to TextInput
