+++ src/js/components/Heading/StyledHeading.js
@@ -1,6 +1,11 @@
 import styled, { css } from 'styled-components';
 
-import { breakpointStyle, genericStyles, normalizeColor } from '../../utils';
+import {
+  breakpointStyle,
+  genericStyles,
+  normalizeColor,
+  textAlignStyle,
+} from '../../utils';
 import { defaultProps } from '../../default-props';
 
 const sizeStyle = props => {
@@ -61,16 +66,6 @@ const fontFamily = props => {
       `
     : '';
 };
-
-const TEXT_ALIGN_MAP = {
-  center: 'center',
-  end: 'right',
-  start: 'left',
-};
-
-const textAlignStyle = css`
-  text-align: ${props => TEXT_ALIGN_MAP[props.textAlign]};
-`;
 
 const truncateStyle = `
   white-space: nowrap;
