+++ src/js/components/Text/StyledText.js
@@ -1,6 +1,6 @@
 import styled, { css } from 'styled-components';
 
-import { genericStyles, normalizeColor } from '../../utils';
+import { genericStyles, normalizeColor, textAlignStyle } from '../../utils';
 import { defaultProps } from '../../default-props';
 
 const sizeStyle = props => {
@@ -17,16 +17,6 @@ const sizeStyle = props => {
     line-height: normal;
   `;
 };
-
-const TEXT_ALIGN_MAP = {
-  center: 'center',
-  end: 'right',
-  start: 'left',
-};
-
-const textAlignStyle = css`
-  text-align: ${props => TEXT_ALIGN_MAP[props.textAlign]};
-`;
 
 const truncateStyle = `
   white-space: nowrap;
