+++ src/js/utils/styles.js
@@ -605,3 +605,13 @@ export const kindPartStyles = (obj, theme, colorValue) => {
   if (obj.extend) styles.push(obj.extend);
   return styles;
 };
+
+const TEXT_ALIGN_MAP = {
+  center: 'center',
+  end: 'right',
+  start: 'left',
+};
+
+export const textAlignStyle = css`
+  text-align: ${props => TEXT_ALIGN_MAP[props.textAlign]};
+`;
